+++ app/lib/Settings.cpp
@@ -111,6 +111,7 @@ Language system_default_language()
         case QLocale::Italian: return Language::Italian;
         case QLocale::Spanish: return Language::Spanish;
         case QLocale::Turkish: return Language::Turkish;
+        case QLocale::Dutch: return Language::Dutch;
         default: return Language::English;
     }
 }